refactor(budget): split alert tests for GH727
